test(ios): fold the trackpad test into caret.ts and drop the containment test
The separate spec existed only to pin a device, which the iOS 17 result showed
was unnecessary. Nothing else justified the split.

The containment test went with it. It swept every [data-editable] asserting
isContentEditable, which is the fix's implementation rather than any behaviour a
user can observe, and it did so at the slowest level available. Per the testing
conventions, an iOS test earns its cost only through a risk lower levels cannot
reach — device behaviour, input mapping, a visual regression — and a DOM
property set by React from Redux state is none of those. The scrub test covers
the same ground through the gesture that actually breaks it.

It was also the least trustworthy test on the branch: green in CI, failing
locally on both iOS versions, and always in its own setup rather than its
assertion.
